reorganized files into separate components and pages folders.  Added prettierignore.
