+++ src/components/formvalidator/CharacterFormValidator.js
